# backend/app/features/pricing/enrichment/router.py
"""Field-value override/revert routes (spec.md route table)."""

import uuid

from fastapi import APIRouter, Depends
from sqlalchemy.ext.asyncio import AsyncSession

from app.core.auth import CurrentStaff, get_scoped_application
from app.core.db import get_db
from app.features.applications.models import Application
from app.features.applications.sections.events import record_field_event
from app.features.pricing.enrichment.schemas import FieldValueOverrideRequest, FieldValueRead
from app.features.pricing.enrichment.service import override_field_value, revert_field_value

router = APIRouter(tags=["pricing"])


@router.patch(
    "/applications/{application_id}/field-values/{field_key}", response_model=FieldValueRead
)
async def patch_field_value(
    application_id: uuid.UUID,
    field_key: str,
    request: FieldValueOverrideRequest,
    user: CurrentStaff,
    db: AsyncSession = Depends(get_db),
    _application: Application = Depends(get_scoped_application),
) -> FieldValueRead:
    row = await override_field_value(db, application_id, field_key, request.value, user.id)
    result = FieldValueRead.model_validate(row)
    # CQ-028a (AC7): activity event naming the field.
    await record_field_event(
        db, application_id, user_id=user.id, reverted=False, field_key=field_key
    )
    return result


@router.post(
    "/applications/{application_id}/field-values/{field_key}/revert",
    response_model=FieldValueRead,
)
async def revert_field_value_route(
    application_id: uuid.UUID,
    field_key: str,
    user: CurrentStaff,
    db: AsyncSession = Depends(get_db),
    _application: Application = Depends(get_scoped_application),
) -> FieldValueRead:
    row = await revert_field_value(db, application_id, field_key)
    result = FieldValueRead.model_validate(row)
    # CQ-028a (AC7): activity event naming the field.
    await record_field_event(
        db, application_id, user_id=user.id, reverted=True, field_key=field_key
    )
    return result