# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clear Quote API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/applications/{application_id}/field-values/{field_key}": {
      "patch": {
        "tags": [
          "pricing"
        ],
        "summary": "Patch Field Value",
        "operationId": "patch_field_value_api_v1_applications__application_id__field_values__field_key__patch",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          },
          {
            "name": "field_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Field Key"
            }
          },
          {
            "name": "cq_staff_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cq Staff Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldValueOverrideRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldValueRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{application_id}/field-values/{field_key}/revert": {
      "post": {
        "tags": [
          "pricing"
        ],
        "summary": "Revert Field Value Route",
        "operationId": "revert_field_value_route_api_v1_applications__application_id__field_values__field_key__revert_post",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          },
          {
            "name": "field_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Field Key"
            }
          },
          {
            "name": "cq_staff_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cq Staff Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldValueRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/portal/applications": {
      "post": {
        "tags": [
          "portal-apply"
        ],
        "summary": "Create Or Get Draft",
        "description": "Creates the borrower's draft, or returns the open one.",
        "operationId": "create_or_get_draft_api_v1_portal_applications_post",
        "parameters": [
          {
            "name": "cq_borrower_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cq Borrower Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyDraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portal/applications/metros": {
      "get": {
        "tags": [
          "portal-apply"
        ],
        "summary": "List Metros",
        "description": "States and their metros for tab 2's picker (decision 28). Declared\nbefore `/{draft_id}` so `metros` is never read as a draft id.",
        "operationId": "list_metros_api_v1_portal_applications_metros_get",
        "parameters": [
          {
            "name": "cq_borrower_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cq Borrower Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetrosOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portal/applications/{draft_id}": {
      "get": {
        "tags": [
          "portal-apply"
        ],
        "summary": "Get Draft",
        "operationId": "get_draft_api_v1_portal_applications__draft_id__get",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          },
          {
            "name": "cq_borrower_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cq Borrower Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyDraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portal/applications/{draft_id}/draft": {
      "patch": {
        "tags": [
          "portal-apply"
        ],
        "summary": "Autosave Tab",
        "description": "Saves one tab and returns its validation errors (never blocks).",
        "operationId": "autosave_tab_api_v1_portal_applications__draft_id__draft_patch",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          },
          {
            "name": "cq_borrower_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cq Borrower Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftPatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portal/applications/{draft_id}/submit": {
      "post": {
        "tags": [
          "portal-apply"
        ],
        "summary": "Submit Application",
        "operationId": "submit_application_api_v1_portal_applications__draft_id__submit_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          },
          {
            "name": "cq_borrower_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cq Borrower Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portal/applications/{draft_id}/documents": {
      "post": {
        "tags": [
          "portal-apply"
        ],
        "summary": "Upload Document",
        "description": "PDF, JPG or PNG, 10 MB at most (AC7). The multipart body is parsed\nby hand, only after auth, ownership and the document cap pass\n(decision 26); `UploadBodyLimitMiddleware` has already bounded its\nsize (411 without `Content-Length`, 413 over the cap).",
        "operationId": "upload_document_api_v1_portal_applications__draft_id__documents_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          },
          {
            "name": "cq_borrower_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cq Borrower Session"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftDocumentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "title": "DraftDocumentUpload",
                "required": [
                  "file",
                  "doc_type"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "title": "File",
                    "description": "PDF, JPG or PNG, 10 MB at most."
                  },
                  "doc_type": {
                    "type": "string",
                    "enum": [
                      "pay_stub",
                      "w2",
                      "bank_statement"
                    ],
                    "title": "Doc Type"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portal/applications/{draft_id}/documents/{document_id}": {
      "delete": {
        "tags": [
          "portal-apply"
        ],
        "summary": "Delete Document",
        "operationId": "delete_document_api_v1_portal_applications__draft_id__documents__document_id__delete",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "cq_borrower_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cq Borrower Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplyDraftOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "current_tab": {
            "$ref": "#/components/schemas/TabName"
          },
          "tabs": {
            "$ref": "#/components/schemas/DraftTabs"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "submitted_application_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted Application Id"
          },
          "consent": {
            "$ref": "#/components/schemas/ConsentTextOut"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "current_tab",
          "tabs",
          "data",
          "submitted_application_id",
          "consent",
          "created_at",
          "updated_at"
        ],
        "title": "ApplyDraftOut"
      },
...
      "ConsentTextOut": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "version",
          "text"
        ],
        "title": "ConsentTextOut"
      },
...
      "DraftDocumentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "doc_type": {
            "type": "string",
            "enum": [
              "pay_stub",
              "w2",
              "bank_statement"
            ],
            "title": "Doc Type"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "doc_type",
          "filename",
          "content_type",
          "size_bytes",
          "uploaded_at"
        ],
        "title": "DraftDocumentOut"
      },
      "DraftPatchRequest": {
        "properties": {
          "tab": {
            "$ref": "#/components/schemas/TabName"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "tab",
          "data"
        ],
        "title": "DraftPatchRequest"
      },
      "DraftPatchResponse": {
        "properties": {
          "draft": {
            "$ref": "#/components/schemas/ApplyDraftOut"
          },
          "tab": {
            "$ref": "#/components/schemas/TabName"
          },
          "tab_valid": {
            "type": "boolean",
            "title": "Tab Valid"
          },
          "field_errors": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Field Errors"
          }
        },
        "type": "object",
        "required": [
          "draft",
          "tab",
          "tab_valid",
          "field_errors"
        ],
        "title": "DraftPatchResponse"
      },
      "DraftTabs": {
        "properties": {
          "you": {
            "$ref": "#/components/schemas/TabStatus"
          },
          "property": {
            "$ref": "#/components/schemas/TabStatus"
          },
          "income": {
            "$ref": "#/components/schemas/TabStatus"
          },
          "consent": {
            "$ref": "#/components/schemas/TabStatus"
          }
        },
        "type": "object",
        "required": [
          "you",
          "property",
          "income",
          "consent"
        ],
        "title": "DraftTabs"
      },
...
      "FieldSource": {
        "type": "string",
        "enum": [
          "encompass",
          "rentcast",
          "airdna",
          "smartasset",
          "steadily",
          "optimal_blue",
          "credit_bureau",
          "property_search",
          "lo_entry",
          "formula",
          "default",
          "lo_override"
        ],
        "title": "FieldSource",
        "description": "Drives the source badge + \"revert to source\" UI on `field_values`."
      },
      "FieldValueOverrideRequest": {
        "properties": {
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "FieldValueOverrideRequest",
        "description": "Every overridable pricing field (spec.md's fixed 5-key list) is a\ncurrency or rate `Decimal`."
      },
      "FieldValueRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "application_id": {
            "type": "string",
            "format": "uuid",
            "title": "Application Id"
          },
          "field_key": {
            "type": "string",
            "title": "Field Key"
          },
          "value": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "source": {
            "$ref": "#/components/schemas/FieldSource"
          },
          "source_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ref"
          },
          "overridden_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overridden By"
          },
          "overridden_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overridden At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "application_id",
          "field_key",
          "value",
          "source",
          "source_ref",
          "overridden_by",
          "overridden_at"
        ],
        "title": "FieldValueRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MetrosOut": {
        "properties": {
          "states": {
            "items": {
              "$ref": "#/components/schemas/StateMetros"
            },
            "type": "array",
            "title": "States"
          }
        },
        "type": "object",
        "required": [
          "states"
        ],
        "title": "MetrosOut",
        "description": "Tab 2's two-tier metro picker: states, each with its metros (both\nsorted). Only these names pass the `buy_box_metros` rule."
      },
...
      "StateMetros": {
        "properties": {
          "state": {
            "type": "string",
            "title": "State"
          },
          "metros": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Metros"
          }
        },
        "type": "object",
        "required": [
          "state",
          "metros"
        ],
        "title": "StateMetros"
      },
...
      "SubmitResponse": {
        "properties": {
          "application_id": {
            "type": "string",
            "format": "uuid",
            "title": "Application Id"
          },
          "draft_id": {
            "type": "string",
            "format": "uuid",
            "title": "Draft Id"
          },
          "status": {
            "type": "string",
            "const": "intake",
            "title": "Status"
          },
          "assigned_lo_name": {
            "type": "string",
            "title": "Assigned Lo Name"
          },
          "pipeline_started": {
            "type": "boolean",
            "title": "Pipeline Started"
          }
        },
        "type": "object",
        "required": [
          "application_id",
          "draft_id",
          "status",
          "assigned_lo_name",
          "pipeline_started"
        ],
        "title": "SubmitResponse"
      },
...
      "TabName": {
        "type": "string",
        "enum": [
          "you",
          "property",
          "income",
          "consent"
        ],
        "title": "TabName"
      },
...
      "TabStatus": {
        "properties": {
          "complete": {
            "type": "boolean",
            "title": "Complete"
          }
        },
        "type": "object",
        "required": [
          "complete"
        ],
        "title": "TabStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}